msm: move BlockType and IdentifyBlockType to block_type.go
Addresses review comment: both are self-contained and not specifically
about the StateMachine wiring. Relocating them out of msm.go keeps that
file focused on the state machine itself.
